+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `RuleGroup::limit()`
 - `RuleGroup::last_evaluation()`
 - `RuleGroup::evaluation_time()`
+- `AlertingRule::last_evaluation()`
+- `AlertingRule::evaluation_time()`
+- `RecordingRule::last_evaluation()`
+- `RecordingRule::evaluation_time()`
+- `Rule::as_recording()`
+- `Rule::as_alerting()`
+- `impl Eq for RuleHealth`
+
 ### Changed
 - Refactored deserialization of Prometheus server responses so the explicit dependency on `serde_json` could be removed.
 - Refactored the `error` module and some custom errors related to deserialization using `serde::de::Error`. The `Error` enum inside the `error` module now contains one variant less and existing error variants were improved by _properly_ implementing `std::error::Error::source()`. Libraries like `anyhow` are now able to display more detailed error messages. This change is not breaking if you did not match on specific error enum variants before.